+++ lib/flame_on/component/capture_schema.ex
@@ -3,6 +3,8 @@ defmodule FlameOn.Component.CaptureSchema do
   import Ecto.Changeset
   alias Ecto.Changeset
 
+  require Logger
+
   schema "capture" do
     field :module, :string
     field :function, :string
@@ -10,9 +12,12 @@ defmodule FlameOn.Component.CaptureSchema do
     field :timeout, :integer
   end
 
-  @default_attrs %{module: "cowboy_handler", function: "execute", arity: 2, timeout: 15000}
+  @default_cowboy_attrs %{module: "cowboy_handler", function: "execute", arity: 2, timeout: 15000}
+  @default_bandit_attrs %{module: "Bandit.Pipeline", function: "run", arity: 4, timeout: 15000}
 
-  def changeset(node, attrs \\ @default_attrs) do
+  def changeset(node, attrs \\ nil) do
+    attrs = attrs || default_attrs(node)
+
     %__MODULE__{}
     |> cast(attrs, [:module, :function, :arity, :timeout])
     |> validate_required([:module, :function, :arity, :timeout])
@@ -85,5 +90,19 @@ defmodule FlameOn.Component.CaptureSchema do
 
   defp rpc_function_exported?(module, function, arity, node) do
     :rpc.call(node, Kernel, :function_exported?, [module, function, arity])
+  end
+
+  defp default_attrs(node) do
+    cond do
+      rpc_function_exported?(@default_bandit_attrs.module, @default_bandit_attrs.function, @default_bandit_attrs.arity, node) ->
+        @default_bandit_attrs
+
+      rpc_function_exported?(@default_cowboy_attrs.module, @default_cowboy_attrs.function, @default_cowboy_attrs.arity, node) ->
+        @default_cowboy_attrs
+
+      true ->
+        Logger.info("FlameOn did not detect Cowboy or Bandit modules")
+        %{}
+    end
   end
 end