+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Next
+
+- `Phoenix.HTML` 4.0
+- Bandit.Pipeline.run as default MFA when available (it's now the default in `mix phx_new`)
+- Fix issue with changeset form not honoring initial node
+- Elixir 1.16 in CI
+- Increase formatter line length
+
 ## 0.6.0
 
 - Upgrade Dialyxir